requirements.txt
addict==2.4.0
adversarial-robustness-toolbox==1.16.0
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.12.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blessed==1.20.0
blinker==1.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorlog==6.7.0
comm==0.2.0
contourpy==1.2.0
crcmod==1.7
cryptography==41.0.5
#ctcdecode==1.0.3
cycler==0.12.1
datasets==2.13.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dnspython==2.4.2
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.1.3
executing==2.0.1
fast-bss-eval==0.1.4
fastapi==0.104.1
fastjsonschema==2.18.1
ffmpy==0.3.1
filelock==3.12.0
flask==3.0.0
fonttools==4.44.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.10
gitpython==3.1.31
gpustat==1.1
gradio==3.48.0
gradio-client==0.6.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.21.4
hydra-configs-pytorch-lightning==0.1.0
hydra-core==1.3.2
hyperpyyaml==1.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter-server==2.10.0
jupyter-server-terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-server==2.25.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy-loader==0.2
levenshtein==0.23.0
librosa==0.9.2
lightning-utilities==0.8.0
lion-pytorch==0.1.2
lit==16.0.2
llvmlite==0.40.0
mako==1.3.0
mamba-ssm==1.2.0.post1
markdown-it-py==2.2.0
markupsafe==2.1.2
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
modelscope==1.9.5
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1.1
notebook==7.0.6
notebook-shim==0.2.3
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python==4.7.0.72
optuna==3.4.0
orjson==3.9.10
oss2==2.18.3
overrides==7.4.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pillow==9.5.0
platformdirs==4.0.0
pluggy==1.3.0
pooch==1.6.0
prometheus-client==0.18.0
prompt-toolkit==3.0.40
protobuf==4.22.3
psutil==5.9.5
ptflops==0.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic-core==2.10.1
pydub==0.25.1
pygments==2.15.1
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-etcd==0.4.5
python-json-logger==2.0.7
python-levenshtein==0.23.0
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytorch-ranger==0.1.1
pytz==2023.3
pyyaml==6.0
pyzmq==25.1.1
qtconsole==5.5.0
qtpy==2.4.1
rapidfuzz==3.5.2
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rpds-py==0.12.0
ruamel-yaml==0.17.22
ruamel-yaml-clib==0.2.7
safetensors==0.4.2
scikit-learn==1.1.3
scipy==1.10.1
semantic-version==2.10.0
send2trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.21.1
setproctitle==1.3.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
sox==1.4.1
soxr==0.3.5
speechbrain==0.5.14
sqlalchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
sympy==1.11.1
terminado==0.18.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torch-complex==0.4.3
torch-mir-eval==0.4
torch-optimizer==0.3.0
torchaudio==2.0.1
torchelastic==0.2.2
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.13.0
transformers==4.38.2
typeguard==2.13.3
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
uvicorn==0.24.0.post1
wandb==0.15.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.9
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0